first-steps: Do not add any more NVIDIA modules to /etc/mkinitcpio.conf
This greatly inflates your boot partition, which often causes you to run out of space.

It also causes errors when upgrading images, like:

module not found: `nvidia(_modeset/_uvm/_drm)`

This is simply not necessary anymore, because in the next section we install package nvidia-tweaks, where there is a ready-made hook to update all the images when you update the drivers/nvidia-tweaks package itself. This is usually enough to avoid black screen/modules dropout.
https://aur.archlinux.org/cgit/aur.git/commit/?h=nvidia-tweaks&id=0da92e01937f8526c94e8f83d402a9d5a7de8323