custom-kernels: Remove liquorix kernel
Oh yes, I will be hated for this.

The reasons for removal are as follows:

- Sucks the kernel config from Debian

- PDS scheduler which has a number of issues

- Lots of problems with AMD processors (this was highlighted in warning)

Finally, linux-cachyos-pds is simply better:

https://flightlessmango.com/games/23255/logs/3772
https://flightlessmango.com/games/13744/logs/3773
https://flightlessmango.com/games/13744/logs/3645

I emphasize, in all my tests linux-lqx - sucks.

Personally, I really loved this kernel before, in the year 2020, but now
you just have to accept that the kernel has lost its relevance as well
as Xanmod. For now, I recommend that everyone use linux-tkg/linux-cachyos.

This commit can be reverted if someone gives me a good enough reason to
return this kernel.