extra-optimizations: Do not rebuild LLVM/Clang manually
It's a waste of time. First, you won't be able to build many packages
via llvm-git because they will require a specific version of LLVM.
Second, it doesn't give you a significant gain in build speed. The most
you can gain is a couple of minutes, but to build a new Clang every time
you need it, you will need many hours on some hardware. It is better to
just install Clang from the CachyOS repositories

Signed-off-by: Vasiliy Stelmachenok <ventureo@cachyos.org>