extra-optimizations: Remove increasing limits
This was usually only needed for Esync to work. But Esync at the moment
looks very outdated and works in a slightly “hacker” way by creating a
bunch of file descriptors, although it may work better for low-end CPUs
with few cores, in the end it doesn't give much gain and enough accuracy
to reproduce NT primitives behavior. The best option right now is to
install a kernel and Wine with NTSync support.

Signed-off-by: Vasiliy Stelmachenok <ventureo@cachyos.org>